+++ include/m3uparser.hpp
@@ -1,5 +1,4 @@
-#ifndef M3UPARSER_HPP
-#define M3UPARSER_HPP
+#pragma once
 
 #include <QString>
 #include <QList>
@@ -12,9 +11,13 @@ struct M3UItem {
 
 class M3UParser {
 public:
+
     M3UParser() = default;
 
+    /**
+     * @brief Analiza el contenido de texto de una lista de reproducción M3U y extrae sus elementos.
+     * @param content Cadena de texto con la estructura de un archivo M3U/M3U8.
+     * @return Lista de estructuras M3UItem obtenidas del contenido.
+     */
     QList<M3UItem> parseContent(const QString& content);
 };
-
-#endif // M3UPARSER_HPP